this will delete all unimportant messages every 6 months
The server run an infinite loop to check what the date is. If it has been 6 months since the start of the server, it will clear all messages, except for the ones that have been flagged as important. For the important messages, it deletes the chatId field, so that the chats table can be cleared